extract init_specs_to_fetch to openshift_base
