+++ reconcile/openshift_namespaces.py
@@ -2,7 +2,7 @@ import logging
 
 import utils.gql as gql
 import utils.threaded as threaded
-import reconcile.openshift_resources as openshift_resources
+import reconcile.openshift_base as ob
 
 from utils.openshift_resource import ResourceInventory
 from utils.oc import OC_Map
@@ -47,7 +47,7 @@ def get_desired_state():
     namespaces = gqlapi.query(QUERY)['namespaces']
     ri = ResourceInventory()
     oc_map = OC_Map(namespaces=namespaces, integration=QONTRACT_INTEGRATION)
-    openshift_resources.init_specs_to_fetch(
+    ob.init_specs_to_fetch(
         ri,
         oc_map,
         namespaces,
