+++ reconcile/openshift_network_policies.py
@@ -4,6 +4,7 @@ import semver
 import utils.gql as gql
 import utils.threaded as threaded
 import reconcile.openshift_resources as openshift_resources
+import reconcile.openshift_base as ob
 
 from utils.openshift_resource import ResourceInventory, OpenshiftResource
 from utils.oc import OC_Map
@@ -86,7 +87,7 @@ def fetch_current_state(namespaces, thread_pool_size):
     ri = ResourceInventory()
     oc_map = OC_Map(namespaces=namespaces, integration=QONTRACT_INTEGRATION)
     state_specs = \
-        openshift_resources.init_specs_to_fetch(
+        ob.init_specs_to_fetch(
             ri,
             oc_map,
             namespaces,
