+++ reconcile/terraform_resources.py
@@ -5,6 +5,7 @@ import semver
 import utils.gql as gql
 import utils.threaded as threaded
 import reconcile.openshift_resources as openshift_resources
+import reconcile.openshift_base as ob
 
 from reconcile.queries import AWS_ACCOUNTS_QUERY
 from utils.terrascript_client import TerrascriptClient as Terrascript
@@ -87,7 +88,7 @@ def fetch_current_state(namespaces, thread_pool_size):
     ri = ResourceInventory()
     oc_map = OC_Map(namespaces=namespaces, integration=QONTRACT_INTEGRATION)
     state_specs = \
-        openshift_resources.init_specs_to_fetch(
+        ob.init_specs_to_fetch(
             ri,
             oc_map,
             namespaces,
